+++ libs/validation/tests/validation/data_transformation/test_formatting.py
@@ -218,7 +218,10 @@ def test_risk_state_person_time(risk_state_person_time_data: pd.DataFrame) -> No
     formatter = RiskStatePersonTime("child_stunting")
 
     assert formatter.measure == "person_time"
-    assert formatter.is_person_time
+    assert (
+        formatter.to_opportunity_counts(pd.DataFrame({"value": [10.0]}), 0.1)["value"].iloc[0]
+        == 100
+    )
     assert formatter.entity == "child_stunting"
     assert formatter.raw_dataset_name == "person_time_child_stunting"
     assert formatter.sum_all == False
@@ -252,7 +255,10 @@ def test_risk_state_person_time_sum_all(risk_state_person_time_data: pd.DataFram
     formatter = RiskStatePersonTime("child_stunting", sum_all=True)
 
     assert formatter.measure == "person_time"
-    assert formatter.is_person_time
+    assert (
+        formatter.to_opportunity_counts(pd.DataFrame({"value": [10.0]}), 0.1)["value"].iloc[0]
+        == 100
+    )
     assert formatter.entity == "child_stunting"
     assert formatter.raw_dataset_name == "person_time_child_stunting"
     assert formatter.sum_all == True